address pr comments and fix deregistration request
