+++ internal/control_test_engine/ue/nas/message/nas_control/mm_5gs/deregistration-request.go
@@ -8,12 +8,23 @@ import (
 	"bytes"
 	"fmt"
 	"my5G-RANTester/internal/control_test_engine/ue/context"
+	"my5G-RANTester/internal/control_test_engine/ue/nas/message/nas_control"
 
 	"github.com/free5gc/nas"
 	"github.com/free5gc/nas/nasMessage"
 )
 
-func GetDeregistrationRequest(ue *context.UEContext) (nasPdu []byte) {
+func DeregistrationRequest(ue *context.UEContext) ([]byte, error) {
+
+	pdu := getDeregistrationRequest(ue)
+	pdu, err := nas_control.EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCipheredWithNew5gNasSecurityContext, true, false)
+	if err != nil {
+		return nil, fmt.Errorf("Error encoding %s IMSI UE  NAS Security Mode Complete message", ue.UeSecurity.Supi)
+	}
+	return pdu, nil
+}
+
+func getDeregistrationRequest(ue *context.UEContext) (nasPdu []byte) {
 	m := nas.NewMessage()
 	m.GmmMessage = nas.NewGmmMessage()
 	m.GmmHeader.SetMessageType(nas.MsgTypeDeregistrationRequestUEOriginatingDeregistration)
