+++ test/aio5gc/context/context.go
@@ -6,12 +6,14 @@ package context
 
 import (
 	"errors"
+	"fmt"
 	"my5G-RANTester/config"
 	"strconv"
 
 	"github.com/free5gc/nas"
 	"github.com/free5gc/ngap/ngapType"
 	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/util/fsm"
 )
 
 // All in one 5GC for test purpose
@@ -31,7 +33,7 @@ func (a *Aio5gc) GetSessionContext() *SessionContext {
 	return &a.session
 }
 
-func (a *Aio5gc) Init(conf config.Config, id string, name string) error {
+func (a *Aio5gc) Init(conf config.Config, id string, name string, ueCallbacks map[fsm.StateType]fsm.Callback, pduCallbacks map[fsm.StateType]fsm.Callback) error {
 	plmn := models.PlmnId{
 		Mcc: conf.GNodeB.PlmnList.Mcc,
 		Mnc: conf.GNodeB.PlmnList.Mnc,
@@ -59,6 +61,9 @@ func (a *Aio5gc) Init(conf config.Config, id string, name string) error {
 		},
 	}
 
+	pdufsm, err := initPduFSM(pduCallbacks)
+	uefsm, err := initUeFSM(ueCallbacks)
+
 	a.conf = conf
 	a.amfContext = AMFContext{}
 	a.amfContext.NewAmfContext(
@@ -67,6 +72,8 @@ func (a *Aio5gc) Init(conf config.Config, id string, name string) error {
 		supportedPlmns,
 		servedGuami,
 		100,
+		uefsm,
+		pdufsm,
 	)
 
 	a.session.NewSessionContext()
@@ -87,4 +94,79 @@ func (a *Aio5gc) GetNgapHooks() []func(*ngapType.NGAPPDU, *GNBContext, *Aio5gc) 
 
 func (a *Aio5gc) SetNgapHooks(hook []func(*ngapType.NGAPPDU, *GNBContext, *Aio5gc) (bool, error)) {
 	a.ngapHook = hook
+}
+
+func initUeFSM(callbacks fsm.Callbacks) (*fsm.FSM, error) {
+
+	if callbacks == nil {
+		callbacks = fsm.Callbacks{}
+	}
+	states := []fsm.StateType{AuthenticationInitiated, Authenticated, Registered, DeregisteredInitiated, Deregistered}
+	for _, i := range states {
+		_, ok := callbacks[i]
+		if !ok {
+			callbacks[i] = func(state *fsm.State, event fsm.EventType, args fsm.ArgsType) {}
+		}
+	}
+
+	UeFsm, err := fsm.NewFSM(
+		fsm.Transitions{
+			{Event: RegistrationRequest, From: Deregistered, To: AuthenticationInitiated},
+			{Event: AuthenticationSuccess, From: AuthenticationInitiated, To: Authenticated},
+			{Event: RegistrationAccept, From: Authenticated, To: Registered},
+			{Event: DeregistrationRequest, From: Registered, To: DeregisteredInitiated},
+			{Event: Deregistration, From: DeregisteredInitiated, To: Deregistered},
+			{Event: ForceDeregistrationInit, From: AuthenticationInitiated, To: DeregisteredInitiated},
+			{Event: ForceDeregistrationInit, From: Authenticated, To: DeregisteredInitiated},
+			{Event: ForceDeregistrationInit, From: Registered, To: DeregisteredInitiated},
+			{Event: ForceDeregistrationInit, From: DeregisteredInitiated, To: DeregisteredInitiated},
+		},
+		fsm.Callbacks{
+			AuthenticationInitiated: callbacks[AuthenticationInitiated],
+			Authenticated:           callbacks[Authenticated],
+			Registered:              callbacks[Registered],
+			DeregisteredInitiated:   callbacks[DeregisteredInitiated],
+			Deregistered:            callbacks[Deregistered],
+		},
+	)
+	if UeFsm == nil || err != nil {
+		return nil, fmt.Errorf("[5GC] Failed to create PDU FSM: %v", err.Error())
+	}
+	return UeFsm, nil
+}
+
+func initPduFSM(callbacks fsm.Callbacks) (*fsm.FSM, error) {
+
+	if callbacks == nil {
+		callbacks = fsm.Callbacks{}
+	}
+	states := []fsm.StateType{Inactive, InactivePending, Active, ModificationPending}
+	for _, i := range states {
+		_, ok := callbacks[i]
+		if !ok {
+			callbacks[i] = func(state *fsm.State, event fsm.EventType, args fsm.ArgsType) {}
+		}
+	}
+
+	PduFsm, err := fsm.NewFSM(
+		fsm.Transitions{
+			{Event: EstablishmentReject, From: Inactive, To: Inactive},
+			{Event: EstablishmentAccept, From: Inactive, To: Active},
+			{Event: ReleaseComplete, From: InactivePending, To: Inactive},
+			{Event: ReleaseCommand, From: Active, To: InactivePending},
+			{Event: ModificationCommand, From: Active, To: ModificationPending},
+			{Event: ModificationComplete, From: ModificationPending, To: Inactive},
+			{Event: ForceRelease, From: Active, To: Inactive},
+		},
+		fsm.Callbacks{
+			Inactive:            callbacks[Inactive],
+			InactivePending:     callbacks[InactivePending],
+			Active:              callbacks[Active],
+			ModificationPending: callbacks[ModificationPending],
+		},
+	)
+	if PduFsm == nil || err != nil {
+		return nil, fmt.Errorf("[5GC] Failed to create PDU FSM: %v", err.Error())
+	}
+	return PduFsm, nil
 }
