+++ test/aio5gc/context/sessionManagement.go
@@ -7,7 +7,6 @@ package context
 import (
 	"errors"
 	"fmt"
-	"my5G-RANTester/test/aio5gc/state"
 	"net"
 
 	"github.com/free5gc/nas/nasConvert"
@@ -41,10 +40,27 @@ type ProtocolConfigurationOptions struct {
 	IPv4LinkMTURequest bool
 }
 
+const (
+	Inactive            fsm.StateType = "Inactive"
+	InactivePending     fsm.StateType = "InactivePending"
+	Active              fsm.StateType = "Active"
+	ModificationPending fsm.StateType = "ModificationPending"
+)
+
+const (
+	EstablishmentReject  fsm.EventType = "EstablishmentReject"
+	EstablishmentAccept  fsm.EventType = "EstablishmentAccept"
+	ReleaseComplete      fsm.EventType = "ReleaseComplete"
+	ReleaseCommand       fsm.EventType = "ReleaseCommand"
+	ForceRelease         fsm.EventType = "ForceRelease"
+	ModificationCommand  fsm.EventType = "ModificationCommand"
+	ModificationComplete fsm.EventType = "ModificationComplete"
+)
+
 func NewSmContext(pduSessionID int32) *SmContext {
 	c := &SmContext{pduSessionID: pduSessionID}
 	c.ProtocolConfigurationOptions = &ProtocolConfigurationOptions{}
-	c.state = fsm.NewState(state.Inactive)
+	c.state = fsm.NewState(Inactive)
 	return c
 }
 
@@ -169,8 +185,7 @@ func CreatePDUSession(sessionRequest *nasMessage.PDUSessionEstablishmentRequest,
 			newSmContext.ProtocolConfigurationOptions.IPv4LinkMTURequest = true
 		}
 	}
-
-	err = state.GetPduFSM().SendEvent(newSmContext.GetState(), state.EstablishmentAccept, fsm.ArgsType{"ue": ue, "sm": newSmContext}, log.NewEntry(log.StandardLogger()))
+	err = ue.GetPduFsm().SendEvent(newSmContext.GetState(), EstablishmentAccept, fsm.ArgsType{"ue": ue, "sm": newSmContext}, log.NewEntry(log.StandardLogger()))
 	if err != nil {
 		return nil, err
 	}
@@ -187,7 +202,7 @@ func ReleasePDUSession(ue *UEContext, pduSessionID int32) (SmContext, error) {
 	if err != nil {
 		return SmContext{}, err
 	}
-	err = state.GetPduFSM().SendEvent(sm.state, state.ReleaseCommand, fsm.ArgsType{"ue": ue, "sm": sm}, log.NewEntry(log.StandardLogger()))
+	err = ue.GetPduFsm().SendEvent(sm.state, ReleaseCommand, fsm.ArgsType{"ue": ue, "sm": sm}, log.NewEntry(log.StandardLogger()))
 	if err != nil {
 		return SmContext{}, err
 	}
@@ -199,7 +214,7 @@ func ConfirmPDUSessionRelease(ue *UEContext, pduSessionID int32) error {
 	if err != nil {
 		return err
 	}
-	state.GetPduFSM().SendEvent(sm.state, state.ReleaseComplete, fsm.ArgsType{"ue": ue, "sm": sm}, log.NewEntry(log.StandardLogger()))
+	ue.GetPduFsm().SendEvent(sm.state, ReleaseComplete, fsm.ArgsType{"ue": ue, "sm": sm}, log.NewEntry(log.StandardLogger()))
 	if err != nil {
 		return err
 	}
@@ -209,10 +224,10 @@ func ConfirmPDUSessionRelease(ue *UEContext, pduSessionID int32) error {
 
 func ForceReleaseAllPDUSession(ue *UEContext) {
 	ue.ExecuteForAllSmContexts(func(smCtx *SmContext) {
-		err := state.GetPduFSM().SendEvent(smCtx.state, state.ForceRelease, fsm.ArgsType{"ue": ue, "sm": smCtx}, log.NewEntry(log.StandardLogger()))
+		err := ue.GetPduFsm().SendEvent(smCtx.state, ForceRelease, fsm.ArgsType{"ue": ue, "sm": smCtx}, log.NewEntry(log.StandardLogger()))
 		if err != nil {
 			log.Error("[5GC] Failed to release pdu session " + fmt.Sprint(smCtx.GetPduSessionId()) + " for ue " + ue.securityContext.msin + ": " + err.Error())
 		}
+		ue.DeleteSmContext(smCtx.GetPduSessionId())
 	})
-	ue.DeleteAllSmContext()
 }
