+++ test/aio5gc/msg/nas/handler/ueOriginatingDeregistration.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"my5G-RANTester/test/aio5gc/context"
 	"my5G-RANTester/test/aio5gc/msg"
-	"my5G-RANTester/test/aio5gc/state"
 
 	"github.com/free5gc/nas"
 	"github.com/free5gc/nas/nasMessage"
@@ -21,18 +20,14 @@ import (
 func UEOriginatingDeregistration(nasReq *nas.Message, amf *context.AMFContext, ue *context.UEContext, gnb *context.GNBContext) error {
 	var err error
 	switch ue.GetState().Current() {
-	case state.AuthenticationInitiated:
+	case context.AuthenticationInitiated:
 		err = DefaultUEOriginatingDeregistration(nasReq, amf, ue, gnb)
-	case state.Deregistrated:
-		err = fmt.Errorf("[5GC][NAS] Unexpected message: received UEOriginatingDeregistration for Deregistrated UE")
-	case state.DeregistratedInitiated:
-		err = fmt.Errorf("[5GC][NAS] Unexpected message: received UEOriginatingDeregistration for DeregistratedInitiated UE")
-	case state.Registred:
+	case context.Registered:
 		err = DefaultUEOriginatingDeregistration(nasReq, amf, ue, gnb)
-	case state.Authenticated:
+	case context.Authenticated:
 		err = DefaultUEOriginatingDeregistration(nasReq, amf, ue, gnb)
 	default:
-		err = fmt.Errorf("Unknown UE state: %v ", ue.GetState().Current())
+		err = fmt.Errorf("[5GC][NAS] Unexpected message: received %s for UEOriginatingDeregistration", ue.GetState().Current())
 	}
 	return err
 }
@@ -42,7 +37,7 @@ func DefaultUEOriginatingDeregistration(nasReq *nas.Message, amf *context.AMFCon
 	deregistrationRequest := nasReq.DeregistrationRequestUEOriginatingDeregistration
 	context.ForceReleaseAllPDUSession(ue)
 
-	err := state.GetUeFsm().SendEvent(ue.GetState(), state.DeregistrationRequest, fsm.ArgsType{"ue": ue}, log.NewEntry(log.StandardLogger()))
+	err := ue.GetUeFsm().SendEvent(ue.GetState(), context.DeregistrationRequest, fsm.ArgsType{"ue": ue}, log.NewEntry(log.StandardLogger()))
 
 	if err != nil {
 		return err
