+++ test/aio5gc/msg/ngap/handler/initialContextSetupResponse.go
@@ -32,7 +32,7 @@ func InitialContextSetupResponse(req *ngapType.InitialContextSetupResponse, fgc 
 		}
 	}
 	if ue != ranUe {
-		return errors.New("[5GC][NGAP] RanUeNgapId does not match the one registred for this UE")
+		return errors.New("[5GC][NGAP] RanUeNgapId does not match the one Registered for this UE")
 	}
 	return nil
 }
