+++ test/aio5gc/msg/ngap/handler/pduSessionResourceSetup.go
@@ -33,7 +33,7 @@ func PDUSessionResourceSetup(req *ngapType.PDUSessionResourceSetupResponse, fgc 
 		}
 	}
 	if ue != ranUe {
-		return errors.New("[5GC][NGAP] RanUeNgapId does not match the one registred for this UE")
+		return errors.New("[5GC][NGAP] RanUeNgapId does not match the one Registered for this UE")
 	}
 	return nil
 }
